Home

PotentialFitting.jl Documentation

PotentialFitting.jl Documentation

Package Features

Functions

Package Features

Define new potentials easily

Fit potential using methods provided by ScikitLearn

Functions

Visualize

PotentialFitting.visualize.min_distance — Method.

min_distance(mpp::MoleculePairPotential, points)

Gives minimum distance of molecules in mpp on given points

source

PotentialFitting.visualize.plot_compare — Method.

plot_compare(points, energy, mpp::MoleculePairPotential...; emax=100, unit="cm^-1", leg=false, size=(800,400), font=font(20))

Compares fitted energy to calculated one.

Arguments

points : array of Cluster where potential is plotted

energy : calculated energy

mpp::MoleculePairPotential... : potentials to be fitted

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

font=font(20) : font size

source

PotentialFitting.visualize.plot_potential — Method.

plot_potential(points, mpp::MoleculePairPotential; emax=100, unit="cm^-1", leg=false, size=(800,400), font=font(20))

Plots potential

Arguments

points : array of Cluster where potential is plotted

mpp::MoleculePairPotential : potential

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

font=font(20) : font size

source

PotentialFitting.visualize.scan_compare — Method.

scan_compare(points,energy, mppe...; emax=100, unit="cm^-1", leg=false, fsize=(800,400))

Use Interact to view potentials ineractively on different points.

Visualization is done on collumn vise.

Arguments

points : array of points, first dimension is displayd while second can be chosen

energy : array of reference energy, first dimension is displayd while second can be chosen

mppe... : MoleculePairPotential which are plotted

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

source

PotentialFitting.visualize.scan_vizualize — Method.

scan_vizualize(points; i=4)

Visualize geometry of points interactively using Interact

Arguments

points : array of points, first dimension is displayd while second can be chosen

i=4 : row index at with visialization is done

source

PotentialFitting.visualize.visualize_point_bio3dview — Method.

visualizepointbio3dview(point::Cluster)

Visualize point using Bio3DView. Can be used wit IJulia and on html

source

PotentialFitting.visualize.visualize_points — Method.

visualize_points(points; stdout=devnull, stderr=devnull, command="vmd")

Visualize using external program.

source

Fit

PotentialFitting.fit.fit_potential! — Method.

fit_potential!(model, mpp::MoleculePairPotential, fdata::FitData)

Fits potential using given model

Arguments

model : ScikitLearn model

mpp::MoleculePairPotential : potential

fdata::FitData : data used in fitting

source

PotentialFitting.fit.predict_potential — Method.

predict_potential(model, mpp::MoleculePairPotential, cluster1, cluster2)

Uses model to predict potential on given cluster points

source

PotentialFitting.fit.predict_potential — Method.

predict_potential(model, mpp::MoleculePairPotential, points)

Uses model to predict potential on given points

source

PotentialFitting.fit.rmsd — Method.

rmsd(points, energy, mpp::MoleculePairPotential; emax=0, unit="cm^-1")

Calculates root mean square error for potential mpp.

Atributes

points : points where potential is tested

energy : referece energy for given points

mpp::MoleculePairPotential : potential

emax=0 : cut points where energy is larger than this

unit="cm^-1" : unit for emax

source

PotentialFitting.fit.setweight_e_less! — Function.

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

source

PotentialFitting.fit.setweight_e_more! — Method.

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

source

PotentialFitting.fit.FitData — Type.

FitData

Structure to help potential parameters fitting.

Fields

variables : variables

E : energy

w : weights

source

PotentialFitting.fit.give_as_potential — Method.

giveaspotential(T, data)

Returns MoleculePairPotential{T} from given data that must have "c1_molecule" and "c2_molecule" fields

source

Potentials

Modules = [PotentialFitting.potentials, PotentialFitting.generaljones]
Order   = [:function, :type, :macro]