PotentialFitting

Visualize

PotentialFitting.visualize.min_distance — Method

min_distance(mpp::MoleculePairPotential, points)

Gives minimum distance of molecules in mpp on given points. Used to help plotting.

PotentialFitting.visualize.plot_compare — Method

plot_compare(points, energy, mpp::MoleculePairPotential...; emax=100, unit="cm^-1", leg=false, size=(800,400), font=font(20))

Compares fitted energy to calculated one.

Arguments

points : array of Cluster where potential is plotted

energy : calculated energy

mpp::MoleculePairPotential... : potentials to be fitted

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

font=font(20) : font size

PotentialFitting.visualize.plot_potential — Method

plot_potential(points, mpp::MoleculePairPotential; emax=100, unit="cm^-1", leg=false, size=(800,400), font=font(20))

Plots potential

Arguments

points : array of Cluster where potential is plotted

mpp::MoleculePairPotential : potential

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

font=font(20) : font size

PotentialFitting.visualize.scan_compare — Method

scan_compare(points,energy, mppe...; emax=100, unit="cm^-1", leg=false, fsize=(800,400))

Use Interact to view potentials ineractively on different points.

Visualization is done on collumn vise.

Arguments

points : array of points, first dimension is displayd while second can be chosen

energy : array of reference energy, first dimension is displayd while second can be chosen

mppe... : MoleculePairPotential which are plotted

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

PotentialFitting.visualize.scan_visualize — Method

scan_vizualize(points; i=4, html=false)

Visualize geometry of points interactively using Interact

Arguments

points : array of points, first dimension is displayd while second can be chosen for index manipulation

i=4 : row index at with visialization is done

html=false : if true give output as string of html

PotentialFitting.visualize.visualize_point_bio3dview — Method

visualizepointbio3dview(point::Cluster; html=false)

Visualize point using Bio3DView. Can be used wit IJulia and on html

PotentialFitting.visualize.visualize_points — Method

visualize_points(points; stdout=devnull, stderr=devnull, command="vmd")

Visualize using external program.

Fit

PotentialFitting.fit.fit_potential! — Method

fit_potential!(model, mpp::MoleculePairPotential, fdata::FitData)

Fits potential using given model

Arguments

model : ScikitLearn model

mpp::MoleculePairPotential : potential

fdata::FitData : data used in fitting

PotentialFitting.fit.predict_potential — Method

predict_potential(model, mpp::MoleculePairPotential, cluster1, cluster2)

Uses model to predict potential on given cluster points

PotentialFitting.fit.predict_potential — Method

predict_potential(model, mpp::MoleculePairPotential, points)

Uses model to predict potential on given points

PotentialFitting.fit.setweight_e_less! — Function

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

PotentialFitting.fit.setweight_e_more! — Method

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

StatsBase.rmsd — Method

rmsd(points, energy, mpp::MoleculePairPotential; emax=0, unit="cm^-1")

Calculates root mean square error for potential mpp.

Atributes

points : points where potential is tested

energy : referece energy for given points

mpp::MoleculePairPotential : potential

emax=0 : cut points where energy is larger than this

unit="cm^-1" : unit for energy

PotentialFitting.fit.FitData — Type

FitData

Structure to help potential parameters fitting.

Fields

variables : variables

E : energy

w : weights

PotentialFitting.fit.give_as_potential — Method

giveaspotential(T, data)

Returns MoleculePairPotential{T} from given data that must have "c1_molecule" and "c2_molecule" fields

Potentials

PotentialFitting.potentials.calculate_potential — Method

calculate_potential(potential::AbstractPotential, cluster::AbstractCluster, indices)

Calculates energy from given potential. Needs to be implemented for each potential.

Arguments

potential::AbstractPairPotential : potential

cluster::AbstractCluster : molecule cluster

indices : indices for atoms that represent potential

PotentialFitting.potentials.get_potential! — Method

get_potential!(potential, terms)

Function that sets potential term after potential has been fitted. That is this function transforms fitted term to potential terms. Needs to be implemented for each potential.

Arguments

potential : used potential

terms : terms from fitting method

PotentialFitting.potentials.potential_variables — Method

potential_variables(mpp::MoleculePairPotential, points)

Used to transform cluster data to fittable form.

Arguments

mpp::MoleculePairPotential : potential to be fitted

points : array for clusters

Returns

Array holding an array for each potential term in mpp.topology that can the be used for fitting.

PotentialFitting.potentials.AbstractClusterPotential — Type

AbstractClusterPotential <: AbstractPotential

Potentials between groups of atoms

PotentialFitting.potentials.AbstractPairPotential — Type

AbstractPairPotential  <: AbstractPotential

Potential that depends only on distance between two atoms

PotentialFitting.potentials.ClusterPotentialTopology — Type

ClusterPotentialTopology{T} <: AbstractClusterPotentialTopology{T}

Container that maps a cluster potential to respective atoms.

Fields

potential::T:

indices::Vector{Vector{Int}}: Stores information of atoms between which the potentilal exist

PotentialFitting.potentials.MoleculePairPotential — Type

MoleculePairPotential{T<:AbstractPairPotential} <: AbstractClusterPotential

Structure to hold potential information between two molecules

PotentialFitting.potentials.PairPotentialTopology — Type

PairPotentialTopology{T} <: AbstractPairPotentialTopology{T}

Container that maps a pair potential to respective atoms.

Fields

indices::Vector{Tuple{Int, Int}}: Stores information of atom pair(s) between which the potentilal exist

potential::T: Potential structure

PotentialFitting.potentials.PairPotentialTopology — Method

(p::PairPotentialTopology)(c1::AbstractCluster, c2::AbstractCluster)

Returns potential for given cluster pair. Index for atom 2 is expected to be in cluster c2.

PotentialFitting.potentials.PairPotentialTopology — Method

(p::PairPotentialTopology)(c::AbstractCluster)

Returns potential for given cluster.

PotentialFitting.potentials.PairPotentialTopology — Method

(p::PairPotentialTopology)(r)

Returns potential for given distance.

PotentialFitting.lennardjones.LennardJones — Type

LennardJones <: AbstractPairPotential

Holds constants for Lennard-Jones potential E = C12/R^12 - C6/R^6

PotentialFitting.generalpowers.GeneralPowers — Type

GeneralPowers

General potential with customizable powers.

Fields

constants::Vector{Float64} : potential constants

powers::Vector{Int} : powers for radius

PotentialCalculation

Atoms

PotentialCalculation.atoms.AtomOnlySymbol — Type

AtomOnlySymbol <: AbstractAtom

Atom that only has symbol

Fields

id::String : hold description of atom

PotentialCalculation.atoms.AtomWithMass — Type

AtomWithMass <: AbstractAtomWithMass

Atom that has symbol and mass

Fields

id::String : hold description of atom

mass::Float64 : mass of atom

PotentialCalculation.molecules.makeidentical! — Method

makeidentical!(Mol::MoleculeIdenticalInformation,x)

Adds information for identical atoms to molecule.

Arguments

Mol::MoleculeIdenticalInformation : moleculet to which identical information is added

x : collection of indices of identical atoms

Throws

Error if x is out of bounds of Mol

PotentialCalculation.molecules.Molecule — Type

Molecule{T<:AbstractAtom} <: AbstractMolecule

Molecule representation

Field

atoms::Vector{T} : Contains information of atoms

PotentialCalculation.molecules.MoleculeIdenticalInformation — Type

Molecule{T<:AbstractAtom} <: AbstractMolecule

Molecule representation with information of identical atoms

Field

atoms::Vector{T} : Contains information of atoms

identical::Identical : identical atom information

Base.push! — Method

Base.push!(Mol::MoleculeIdenticalInformation,x)

Adds information for identical atoms to molecule.

DEPRECATED!!! - use makeidentical! instead

Arguments

Mol::MoleculeIdenticalInformation : moleculet to which identical information is added

x : collection of indices of identical atoms

Throws

Error if x is out of bounds of Mol

PotentialCalculation.identical.areidentical — Method

areidentical(mol::MoleculeIdenticalInformation,x)

Checks if atoms with indices given in collection x are identical

Calculation

PotentialCalculation.calculators.bsse_corrected_energy — Method

bsse_corrected_energy(cal::Calculator, c1, c2; basename="base", id="", pchannel=pchannel)

Calculates energy of combined clusters taking into account basis set superposition error

Arguments

cal::Calculator : calcualation information

c1 : collection of clusters

c2 : collection of clusters

basename="base" : base name for input/ouput files

id="" : additional information for calculator - needed for multiprocessing in same folder

pchannel=undef : (Remote)Channel where progess information is added

PotentialCalculation.calculators.bsse_corrected_energy — Method

bsse_corrected_energy(cal::Calculator, c1::Cluster, c2::Cluster; basename="base", id="", pchannel=pchannel)

Calculates energy of combined cluster taking into account basis set superposition error

Arguments

cal::Calculator : calcualation information

c1::Cluster : cluster

c2::Cluster : cluster

basename="base" : base name for input/ouput files

id="" : additional information for calculator - needed for multiprocessing in same folder

pchannel=undef : (Remote)Channel where progess information is added

PotentialCalculation.calculators.calculate_energy — Method

calculate_energy(cal::Calculator, points; basename="base", ghost=undef, id="", pchannel=pchannel)

Calculates energy for given clusters.

Arguments

cal::Calculator : calculation information

points : collection of Cluster which energy is calculated

basename="base" : base name for input/ouput files

ghost=undef : indices for atoms considered as ghost

id="" : additional information for calculator - needed for multiprocessing in same folder

pchannel=undef : (Remote)Channel where progess information is added

PotentialCalculation.calculators.calculate_energy — Method

calculate_energy(cal::Calculator, point::Cluster; basename="base", ghost=undef, id="", pchannel=pchannel)

Calculates energy for given cluster.

Arguments

cal::Calculator : calculation information

point::Cluster : cluster which energy is calculated

basename="base" : base name for input/ouput files

ghost=undef : indices for atoms considered as ghost

id="" : additional information for calculator - needed for multiprocessing in same folder

pchannel=undef : (Remote)Channel where progess information is added

PotentialCalculation.calculators.clean_calculation_files — Method

clean_calculation_files(;dir=".", basename="base")

Deletes files that calculations produced

Arguments

dir="." : directory to be cleaned

basename="base" : base name used in calculations

PotentialCalculation.calculators.getBSSEsteps — Method

getBSSEsteps(cal::AbstractCalculationProgram)

Used to set up progressbar. Returns 1, if calculator program has a method to calculate counter poise correction with a single command and 3, if program needs to be called multiple times to make counter poise correction.

PotentialCalculation.calculators.read_energy — Method

read_energy(fname)

Read ORCA output for final energy. Throws an error if calculation failed.

PotentialCalculation.calculators.write_input — Method

write_input(io::IO, cal::Calculator,
                 c::AbstractClusterWithSymbols; ghost=undef)

Writes input files for calculation program. Only ORCA input is supported now.

Arguments

io::IO : stream where writing is done

cal::Calculator : calculation information

c::AbstractClusterWithSymbols : molecule cluster that is calculated

ghost=undef : collection for atom idexes that are considered ghost atoms

PotentialCalculation.calculators.Calculator — Type

mutable struct Calculator

Struct used to hold together different bits of information needed in calculations.

Fields

method : calculation method informaton

basis : basis information

calculator : calculation program information

Examples

julia> Calculator{Orca}("method", "basis")
Calculator{Orca}("method", "basis", Orca("orca", 0x0000000000000001, 0x00000000000003e8, "/tmp/tmpxEoJYW"))

julia> Calculator{Orca}("method", "basis", Orca())
Calculator{Orca}("method", "basis", Orca("orca", 0x0000000000000001, 0x00000000000003e8, "/tmp/tmpVg959k"))

PotentialCalculation.calculators.Orca — Type

mutable struct Orca <: AbstractCalculationProgram

Holds information of how to use ORCA program.

Note that if you are using more than 1 core you need to specify exact path to orca executable.

Fields

executable : path to "orca" executable - default = "orca"

ncore : number of cores used by orca - default = 1

maxmem : maximum memory per core for orca - default = 1000

tmp_dir : directory where calculations are done - default = mktempdir()

PotentialCalculation.calculators.print_orca_xyz — Method

printorcaxyz(io::IO, c::AbstractClusterWithSymbols; ghost=undef)

Prints ORCA input style xyz input. Idexes in ghost are marked as ghost atoms.

PotentialCalculation.restarttools.calculate_adaptive_sample_inputs — Method

calculate_adaptive_sample_inputs(inputs; save_file_name="", save_step=nworkers(), pbar=true)

Uses adaptive_line_sampler to inputs in distributed fashion

Arguments

inputs : calculation inputs array (eg. from load_clusters_and_sample_input)

save_file_name : file where data saved

save_step=nworkers() : number of calculated items before data is saved

pbar=true : show progress bar

PotentialCalculation.restarttools.calculate_energy_for_xyzfile — Method

calculate_energy_for_xyzfile(fname, cal; pbar=true)

Reads xyz-file and calculates energy of each point on it

PotentialCalculation.restarttools.calculate_with_different_method — Method

calculatewithdifferentmethod(fname, calculator::Calculator; savefile="", restartfile="", pbar=true, saveafter=nworkers())

With this function you can use already chosen points on to which to do energy calculation.

Arguments

fname : name of save/restart file where points are read

calculatror::Calculator : calculator used for calculations

save_file="" : save final results here, if not empty

restart_file="" : save restarts here, if not empty

pbar=true : show progress bar

save_after=nworkers() : make restart file when given ammount of points is calculated

PotentialCalculation.restarttools.continue_calculation — Method

continuecalculation(fname, calculator::Calculator; savefile="", restartfile="", saveafter=nworkers(), pbar=true)

Restarts calculation from given file

Arguments

fname : file from which calculation is restarted

calculator::Calculator : calculator used in calculations

save_file : file where final results are saved, if given

restart_file : file where new restart information is saved, if given

save_after=nworkers() : make restart file when given ammount of points is calculated

pbar=true : show progress bar

PotentialCalculation.restarttools.load_clusters_and_make_input — Method

loadclustersandmakeinput(cluster1::Cluster, cluster2::Cluster, calculator; nlines=1, max_e=0, unit="cm-1", npoints=10, maxdis=9.0, sstep=0.1, startdistance=2.5)

Loads cluster1 from xyz-file and returns them all as an Array that can then be used with calculate_adaptive_sample_inputs to calculate energy data. Differs from load_clusters_and_sample_input by taking every point from file

Arguments

cluster1::Cluster : cluster1

cluster2::Cluster : cluster2

nlines : number of lines to be sampled in calculation

max_e : Point that is closest and has less energy than this will be starting point for a line

unit : Unit in which max_e is given

npoints : Number of points in potential

maxdis : Maximum distance in potential calculation

sstep : Search step size for adaptive search

startdistance : Distance from which adaptive seach is started

PotentialCalculation.restarttools.load_clusters_and_make_input — Method

loadclustersandmakeinput(cluster1::String, cluster2::Cluster, calculator, nsamples; max_e=0, unit="cm-1", npoints=10, maxdis=9.0, sstep=0.1, startdistance=2.5)

Loads cluster1 from xyz-file and returns them all as an Array that can then be used with calculate_adaptive_sample_inputs to calculate energy data. Differs from load_clusters_and_sample_input by taking every point from file

Arguments

cluster1::String : file from where cluster1 is sampled

cluster2::Cluster : cluster2

max_e : Point that is closest and has less energy than this will be starting point for a line

unit : Unit in which max_e is given

npoints : Number of points in potential

maxdis : Maximum distance in potential calculation

sstep : Search step size for adaptive search

startdistance : Distance from which adaptive seach is started

PotentialCalculation.restarttools.load_clusters_and_sample_input — Method

loadclustersandsampleinput(fnamecluster1, cluster2, calculator, nsamples; maxe=0, unit="cm-1", npoints=10, maxdis=9.0, sstep=0.1, startdistance=2.5)

Loads cluster1 from xyz-file and takes nsamples samples of it and returns them as an Array that can then be used with calculate_adaptive_sample_inputs to calculate energy data

Arguments

cluster1 : file from where cluster1 is sampled

cluster2 : cluster2

nsamples : number of samples taken from fname_cluster1

max_e : Point that is closest and has less energy than this will be starting point for a line

unit : Unit in which max_e is given

npoints : Number of points in potential

maxdis : Maximum distance in potential calculation

sstep : Search step size for adaptive search

startdistance : Distance from which adaptive seach is started

PotentialCalculation.restarttools.load_clusters_and_sample_input — Method

function loadclustersandsampleinput(cluster1::String, cluster2::String, calculator, nsamples; nlines=1, max_e=0, unit="cm-1", npoints=10, maxdis=9.0, sstep=0.1, startdistance=2.5)

Arguments

cluster1 : file from where cluster1 is sampled

cluster2 : file from where cluster2 is sampled

nsamples : number of samples taken from fname_cluster1

nlines : number of lines to be sampled in calculation

max_e : Point that is closest and has less energy than this will be starting point for a line

unit : Unit in which max_e is given

npoints : Number of points in potential

maxdis : Maximum distance in potential calculation

sstep : Search step size for adaptive search

startdistance : Distance from which adaptive seach is started

PotentialCalculation.restarttools.load_data_file — Method

load_data_file(fname)

Loads saved data

PotentialCalculation.restarttools.load_restart_file — Method

load_restart_file(fname)

Loads restart information from file fname and adds to it key not_calculated, which holds information of which collumns of Points have not been calculated.

PotentialCalculation.restarttools.write_restart_file — Method

write_restart_file(fname, calculator, points, restart_energy, cluster1, cluster2)

Saves restart information for energy calculation.

Arguments

fname : name of restartfile

calculator::Calculator : calculator used in calculations

points : 2d array of point where to calculate energy

energy : energy for points that have been caculeted

cluster1 : cluster1

cluster2 : cluster2

PotentialCalculation.restarttools.write_save_file — Method

write_save_file(fname, calculator, points, energy, cluster1, cluster2)

Saves final information for energy calculation.

Arguments

fname : name of restartfileusing PotentialCalculation

calculator::Calculator : calculator used in calculations

points : 2d array of point where to calculate energy

energy : energy for points that have been caculeted

cluster1 : cluster1

cluster2 : cluster2

PotentialCalculation.restarttools._calculate_points — Method

_calculate_points(cal, c1_points, c2_points; pchannel=undef)

Internal helper function to ease use of distributed calculations. Users should call "calculate_points" instead.

PotentialCalculation.restarttools._sample_and_calculate — Method

_sample_and_calculate(inputs; pchannel=undef )

Internal helper function to help implement distributed calculations

Clusters

PotentialCalculation.clusters.center_cluster! — Method

center_cluster!(c::AbstractCluster)

Centers cluster to origin of coordinates

PotentialCalculation.clusters.center_coordinates — Method

center_coordinates(c::AbstractCluster)

Gives coordinates to aricmetric mean of clusters atoms

PotentialCalculation.clusters.cluster_angle — Method

cluster_angle(c1::AbstractCluster, i, j, c2::AbstractCluster, k)

Calculates angle (radians) between atons in different clusters

Arguments

c1::AbstractCluster : first cluster

i : index in c1

j : index in c1

c2::AbstractCluster : second cluster

k : index in c2

PotentialCalculation.clusters.cluster_angle — Method

cluster_angle(c::AbstractCluster, i, j, k)

Calculates angle (radians) between atoms i,j,k in cluster

PotentialCalculation.clusters.dihedral_angle — Method

clusterdihedralangle(c::AbstractCluster, i, j, k, m)

PotentialCalculation.clusters.distances — Method

distances(c1::AbstractCluster, c2::AbstractCluster)

Return distances between atoms of given clusters

PotentialCalculation.clusters.distances — Method

distances(c1::AbstractCluster, ur1::AbstractUnitRange,
          c2::AbstractCluster, ur2::AbstractUnitRange)

Return distance between atoms ur1 in c1 and atoms ur2 in c2

PotentialCalculation.clusters.distances — Method

distances(c::AbstractCluster, ar1::AbstractRange, ar2::AbstractRange)

Returns distances between atoms in given unit ranges

PotentialCalculation.clusters.distances — Method

distances(c1::AbstractCluster, i, c2::AbstractCluster, j)

Return distance between atom i in c1 and atom j in c2

PotentialCalculation.clusters.distances — Method

distances(c::AbstractCluster, i, j)

Returns distance of atoms i and j

PotentialCalculation.clusters.move! — Method

move!(c::AbstractCluster,r)

Moves cluster by r

PotentialCalculation.clusters.print_xyz — Function

print_xyz(io::IO, c::AbstractClusterWithSymbols, note=""; printheader=true)

Prints cluster in xyz file format

Arguments

io::IO : stream where writing is done

c::AbstractClusterWithSymbols : cluster that is writen

note="" : message writen on note line

printheader=true : wheather or not header is writen (number of atoms and note)

PotentialCalculation.clusters.rotate_x! — Method

rotate_x!(c::AbstractCluster, θ)

Rotates cluster around x-axis by angle θ

PotentialCalculation.clusters.rotate_y! — Method

rotate_y!(c::AbstractCluster, θ)

Rotates cluster around y-axis by angle θ

PotentialCalculation.clusters.rotate_z! — Method

rotate_z!(c::AbstractCluster, θ)

Rotates cluster around z-axis by angle θ

PotentialCalculation.clusters.Cluster — Type

Cluster{T<:AbstractAtom} <: AbstractClusterWithSymbols

Structure to hold location data of clusters/molecules

Fields

xyz::Array{Float64,2} : location of atoms in 3d space, first index is x, y, z coordinate

atoms::Vector{T} : atom type information

PotentialCalculation.clusters.ClusterNoSymbols — Type

ClusterNoSymbols <: AbstractCluster

Basic cluster has only location of atoms.

Fields

xyz::Array{Float64,2} : location of atoms in 3d space, first index is x, y, z coordinate