Home

PotentialFitting.jl Documentation

PotentialFitting.jl Documentation

Package Features

Functions

Usage

Setting up Molecules

Potential Topology

Preparing Data for Fitting

Fitting Potential

Inspecting Fitted Potential

Package Features

Define new potentials easily

Fit potential using methods provided by ScikitLearn

Functions

Visualize

PotentialFitting.visualize.min_distance — Method.

min_distance(mpp::MoleculePairPotential, points)

Gives minimum distance of molecules in mpp on given points. Used to help plotting.

source

PotentialFitting.visualize.plot_compare — Method.

plot_compare(points, energy, mpp::MoleculePairPotential...; emax=100, unit="cm^-1", leg=false, size=(800,400), font=font(20))

Compares fitted energy to calculated one.

Arguments

points : array of Cluster where potential is plotted

energy : calculated energy

mpp::MoleculePairPotential... : potentials to be fitted

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

font=font(20) : font size

source

PotentialFitting.visualize.plot_potential — Method.

plot_potential(points, mpp::MoleculePairPotential; emax=100, unit="cm^-1", leg=false, size=(800,400), font=font(20))

Plots potential

Arguments

points : array of Cluster where potential is plotted

mpp::MoleculePairPotential : potential

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

font=font(20) : font size

source

PotentialFitting.visualize.scan_compare — Method.

scan_compare(points,energy, mppe...; emax=100, unit="cm^-1", leg=false, fsize=(800,400))

Use Interact to view potentials ineractively on different points.

Visualization is done on collumn vise.

Arguments

points : array of points, first dimension is displayd while second can be chosen

energy : array of reference energy, first dimension is displayd while second can be chosen

mppe... : MoleculePairPotential which are plotted

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

source

PotentialFitting.visualize.scan_vizualize — Method.

scan_vizualize(points; i=4)

Visualize geometry of points interactively using Interact

Arguments

points : array of points, first dimension is displayd while second can be chosen

i=4 : row index at with visialization is done

source

PotentialFitting.visualize.visualize_point_bio3dview — Method.

visualizepointbio3dview(point::Cluster)

Visualize point using Bio3DView. Can be used wit IJulia and on html

source

PotentialFitting.visualize.visualize_points — Method.

visualize_points(points; stdout=devnull, stderr=devnull, command="vmd")

Visualize using external program.

source

Fit

PotentialFitting.fit.fit_potential! — Method.

fit_potential!(model, mpp::MoleculePairPotential, fdata::FitData)

Fits potential using given model

Arguments

model : ScikitLearn model

mpp::MoleculePairPotential : potential

fdata::FitData : data used in fitting

source

PotentialFitting.fit.predict_potential — Method.

predict_potential(model, mpp::MoleculePairPotential, cluster1, cluster2)

Uses model to predict potential on given cluster points

source

PotentialFitting.fit.predict_potential — Method.

predict_potential(model, mpp::MoleculePairPotential, points)

Uses model to predict potential on given points

source

PotentialFitting.fit.rmsd — Method.

rmsd(points, energy, mpp::MoleculePairPotential; emax=0, unit="cm^-1")

Calculates root mean square error for potential mpp.

Atributes

points : points where potential is tested

energy : referece energy for given points

mpp::MoleculePairPotential : potential

emax=0 : cut points where energy is larger than this

unit="cm^-1" : unit for emax

source

PotentialFitting.fit.setweight_e_less! — Function.

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

source

PotentialFitting.fit.setweight_e_more! — Method.

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

source

PotentialFitting.fit.FitData — Type.

FitData

Structure to help potential parameters fitting.

Fields

variables : variables

E : energy

w : weights

source

PotentialFitting.fit.give_as_potential — Method.

giveaspotential(T, data)

Returns MoleculePairPotential{T} from given data that must have "c1_molecule" and "c2_molecule" fields

source

Potentials

PotentialFitting.potentials.calculate_potential — Method.

calculate_potential(cluster1::Cluster, cluster2::Cluster, potential, indics::PairTopologyIndices)

Calculates energy from given potential. Needs to be implemented for each potential.

Arguments

cluster1::Cluster : first molecule

cluster2::Cluster : second molecule

potential : potential

indics::PairTopologyIndices : indices for atoms that represent potential

source

PotentialFitting.potentials.calculate_potential — Method.

calculate_potential(mpp::MoleculePairPotential, points)

Calculates potential energy for given points. Returns them as an array of same size as points

source

PotentialFitting.potentials.calculate_potential — Method.

calculate_potential(mpp::MoleculePairPotential, cluster1::Cluster, cluster2::Cluster)

Calculates potential energy

source

PotentialFitting.potentials.get_potential! — Method.

get_potential!(potential, terms)

Function that sets potential term after potential has been fitted. That is this function transforms fitted term to potential terms. Needs to be implemented for each potential.

Arguments

potential : used potential

terms : terms from fitting method

source

PotentialFitting.potentials.potential_variables — Method.

potential_variables(mpp::MoleculePairPotential, cluster1, cluster2)

Used to transform cluster data to fittable form.

Arguments

mpp::MoleculePairPotential : potential to be fitted

cluster1 : array for clusters

cluster2 : array for clusters

Returns

Array holding an array for each potential term in mpp.topology that can the be used for fitting.

source

PotentialFitting.potentials.potential_variables — Method.

potential_variables(mpp::MoleculePairPotential, points)

Used to transform cluster data to fittable form.

Arguments

mpp::MoleculePairPotential : potential to be fitted

cluster1 : array for clusters

Returns

Array holding an array for each potential term in mpp.topology that can the be used for fitting.

source

PotentialFitting.potentials.AbstractClusterPotential — Type.

AbstractClusterPotential <: AbstractPotential

Potentials between groups of atoms

source

PotentialFitting.potentials.AbstractPairPotential — Type.

AbstractPairPotential  <: AbstractPotential

Potential that depends only on distance between two atoms

source

PotentialFitting.potentials.MoleculePairPotential — Type.

MoleculePairPotential{T<:AbstractPairPotential} <: AbstractClusterPotential

Structure to hold potential information between two molecules

source

PotentialFitting.potentials.PairPotentialTopology — Type.

PairPotentialTopology{T}

Potential topology for pair of molecules. Complite potential can be described by several on these structures.

Fields

potential::T : potential

indices::Vector{PairTopologyIndices} : indices for atoms that have this potential

source

PotentialFitting.potentials.PairTopologyIndices — Type.

PairTopologyIndices

Hold information of atom indices for potential calculations.

Fields

first::Vector{Int64} : indices for first molecule

second::Vector{Int64} : indices for second molecule

source

PotentialFitting.potentials.clusters_to_potential_variables — Method.

clusterstopotential_variables(potential, c1::Cluster, c2::Cluster, indeces::PairTopologyIndices)

Returns data in a form that can be used to fit potential (linearly). That is as an array of terms. Example for Lennard Jones potential this is [r^-6 r^-12]. Needs to be implemented for each potential.

source

PotentialFitting.lennardjones.LJ — Type.

LJ

Acronym for LennardJones{Float64}

source

PotentialFitting.lennardjones.LennardJones — Type.

LennardJones{T} <: AbstractPairPotential

Holds constants for Lennard-Jones potential E = C12/R^12 - C6/R^6

source

PotentialFitting.generalpowers.GeneralPowers — Type.

GeneralPowers

General potential with customizable powers.

Fields

constants::Vector{Float64} : potential constants

powers::Vector{Int} : powers for radius

source

PotentialFitting.generalpowers.@GeneralPowers — Macro.

@GeneralPowers(indices, i...)

Used to simplify potential generation. Generates GeneralPowers potential and topology for it.

Examples

julia> @GeneralPowers (1,1) -6 -12
PairPotentialTopology{GeneralPowers}(C(-6)=0.0  C(-12)=0.0  , PairTopologyIndices[PairTopologyIndices([1], [1])])

source