PotentialFitting

Reference

PotentialFitting.calculate_potential — Method

calculate_potential(potential::AbstractPotential, cluster::AbstractCluster, indices)

Calculates energy from given potential. Needs to be implemented for each potential.

Arguments

potential::AbstractPairPotential : potential

cluster::AbstractCluster : molecule cluster

indices : indices for atoms that represent potential

PotentialFitting.fit_potential! — Method

fit_potential!(model, mpp::MoleculePairPotential, fdata::FitData)

Fits potential using given model

Arguments

model : ScikitLearn model

mpp::MoleculePairPotential : potential

fdata::FitData : data used in fitting

PotentialFitting.get_potential! — Method

get_potential!(potential, terms)

Function that sets potential term after potential has been fitted. That is this function transforms fitted term to potential terms. Needs to be implemented for each potential.

Arguments

potential : used potential

terms : terms from fitting method

PotentialFitting.min_distance — Method

min_distance(points, c1::AbstractCluster, c2::AbstractCluster)

Minimum without needing fitted potential.

PotentialFitting.min_distance — Method

min_distance(mpp::MoleculePairPotential, points)

Gives minimum distance of molecules in mpp on given points. Used to help plotting.

PotentialFitting.plot_compare — Method

plot_compare(points, energy, mpp::MoleculePairPotential...; emax=100, unit="cm^-1",
                  leg=false, size=(800,400), font=font(20))

Compares fitted energy to calculated one.

Arguments

points : array of Cluster where potential is plotted

energy : calculated energy

mpp::MoleculePairPotential... : potentials to be fitted

Keywords

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

size=(800,400) : size of picture

font=font(20) : font size

PotentialFitting.plot_potential — Method

plot_potential(points, mpp::MoleculePairPotential; emax=100, unit="cm^-1",
                    leg=false, figsize=(800,400), font=font(20))

Plots potential

Arguments

points : array of Cluster where potential is plotted

mpp::MoleculePairPotential : potential

Keywords

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

figsize=(800,400) : size of picture

font=font(20) : font size

PotentialFitting.plot_potential — Method

plot_potential(pes::Dict; emax=100, unit="cm^-1", figsize=(800,400), font=font(20))

Plots calculated potential

Arguments

pes::Dict : PES in Dict from that can be get with load_data_file command

Keywords

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

figsize=(800,400) : size of picture

font=font(20) : font size

PotentialFitting.potential_variables — Method

potential_variables(mpp::MoleculePairPotential, points)

Used to transform cluster data to fittable form.

Arguments

mpp::MoleculePairPotential : potential to be fitted

points : array for clusters

Returns

Array holding an array for each potential term in mpp.topology that can the be used for fitting.

PotentialFitting.predict_potential — Method

predict_potential(model, mpp::MoleculePairPotential, cluster1, cluster2)

Uses model to predict potential on given cluster points

PotentialFitting.predict_potential — Method

predict_potential(model, mpp::MoleculePairPotential, points)

Uses model to predict potential on given points

PotentialFitting.scan_compare — Method

scan_compare(points,energy, mppe...; emax=100, unit="cm^-1", leg=false, figsize=(800,400))

Use Interact to view potentials ineractively on different points.

Visualization is done on collumn vise.

Arguments

points : array of points, first dimension is displayd while second can be chosen

energy : array of reference energy, first dimension is displayd while second can be chosen

mppe... : MoleculePairPotential which are plotted

Keywords

emax=100 : maximum energy in plot - cut all values with energy greater

unit="cm^-1" : energy unit

leg=false : draw legend

figsize=(800,400) : size of picture

PotentialFitting.setweight_e_less! — Function

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

PotentialFitting.setweight_e_more! — Method

setweightemore!(data::FitData, w, e; unit="cm-1")

Sets weigth when energy is more than given one.

Arguments

data::FitData : data where weigth is adjusted

w : new weigth

e : energy

unit="cm-1" : energy unit

PotentialFitting.visualize_points — Method

visualize_points(points; stdout=devnull, stderr=devnull, command="vmd")

Visualize using external program.

StatsBase.rmsd — Method

rmsd(points, energy, mpp::MoleculePairPotential; emax=0, unit="cm^-1")

Calculates root mean square error for potential mpp.

Atributes

points : points where potential is tested

energy : referece energy for given points

mpp::MoleculePairPotential : potential

emax=0 : cut points where energy is larger than this

unit="cm^-1" : unit for energy

PotentialFitting.AbstractClusterPotential — Type

AbstractClusterPotential <: AbstractPotential

Potentials between groups of atoms

PotentialFitting.AbstractPairPotential — Type

AbstractPairPotential  <: AbstractPotential

Potential that depends only on distance between two atoms

PotentialFitting.ClusterPotentialTopology — Type

ClusterPotentialTopology{T} <: AbstractClusterPotentialTopology{T}

Container that maps a cluster potential to respective atoms.

Fields

potential::T:

indices::Vector{Vector{Int}}: Stores information of atoms between which the potentilal exist

PotentialFitting.FitData — Type

FitData

Structure to help potential parameters fitting.

Fields

variables : variables

E : energy

w : weights

PotentialFitting.GeneralAngle — Type

GeneralAngle <: AbstractClusterPotential

General potential with customizable powers for both radial polynomials and cosθ.

Fields

constants::Array{Float64,2} : constants

ppowers::Vector{Int} : powers for radial polynomials

cpowers::Vector{Int} : powers for cosθ

To create potential use:

GeneralAngle(ppowers::AbstractVector{Int}, cpowers::AbstractVector{Int})

PotentialFitting.GeneralPowers — Type

GeneralPowers

General potential with customizable powers.

Fields

constants::Vector{Float64} : potential constants

powers::Vector{Int} : powers for radius

PotentialFitting.LennardJones — Type

LennardJones <: AbstractPairPotential

Holds constants for Lennard-Jones potential E = C12/R^12 - C6/R^6

PotentialFitting.MoleculePairPotential — Type

MoleculePairPotential{T<:AbstractPairPotential} <: AbstractClusterPotential

Structure to hold potential information between two molecules

PotentialFitting.PairPotentialTopology — Type

PairPotentialTopology{T} <: AbstractPairPotentialTopology{T}

Container that maps a pair potential to respective atoms.

Fields

indices::Vector{Tuple{Int, Int}}: Stores information of atom pair(s) between which the potentilal exist

potential::T: Potential structure

PotentialFitting.PairPotentialTopology — Method

(p::PairPotentialTopology)(c1::AbstractCluster, c2::AbstractCluster)

Returns potential for given cluster pair. Index for atom 2 is expected to be in cluster c2.

PotentialFitting.PairPotentialTopology — Method

(p::PairPotentialTopology)(c::AbstractCluster)

Returns potential for given cluster.

PotentialFitting.PairPotentialTopology — Method

(p::PairPotentialTopology)(r)

Returns potential for given distance.

PotentialFitting.give_as_potential — Method

giveaspotential(T, data)

Returns MoleculePairPotential{T} from given data that must have "c1_molecule" and "c2_molecule" fields